+++ .cypress-cucumber-preprocessorrc.json
@@ -1,7 +1,7 @@
 {
   "stepDefinitions": [
     "cypress/test/[filepath].steps.[jt]s",
-    "cypress/test/[filepart]/*[cC]ommon/**/*.[jt]s"
+    "cypress/test/[filepart]/common/**/*.[jt]s"
   ],
   "messages": {
     "enabled": true,
