+++ cypress/support/commands.ts
@@ -1,7 +1,4 @@
 /// <reference types="cypress" />
-
-import loginPage from "../pages/login.pom";
-
 // ***********************************************
 // This example commands.ts shows you how to
 // create various custom commands and overwrite
@@ -14,36 +11,7 @@ import loginPage from "../pages/login.pom";
 //
 //
 // -- This is a parent command --
-Cypress.Commands.add("interceptApi", (api, method = "GET", times = 1) => {
-  cy.intercept(api.urlRegex, { method, times }, (req) => {
-    delete req.headers["if-none-match"];
-  }).as(api.interceptorName);
-
-  // Make the custom command appear in cypress command log of cypress open mode
-  Cypress.log({
-    message: `Interceptor name ${api.interceptorName} with method ${method} for ${times} times`,
-  });
-});
-
-Cypress.Commands.add("loginWithUI", (user) => {
-  loginPage.typeCredentials(user);
-  loginPage.form.loginBtn().click();
-
-  // Make the custom command appear in cypress command log of cypress open mode
-  Cypress.log({
-    message: `UI login with user ${user}`,
-  });
-});
-
-Cypress.Commands.add("loginProgrammatically", (user) => {
-  // TODO
-
-  // Make the custom command appear in cypress command log of cypress open mode
-  Cypress.log({
-    message: `Progrmmatic login with user ${user}`,
-  });
-});
-
+// Cypress.Commands.add('login', (email, password) => { ... })
 //
 //
 // -- This is a child command --
