+++ cypress/support/consts.ts
@@ -1,31 +1,9 @@
 import { credentials } from "./cyEnvVar";
-import { ApiInfo } from "./models/apiInfo";
 import { UserInfo } from "./models/userInfo";
 
 // Comment const type to allow syntax error upon accessing non-exiting property
 export const users /*: { [key: string]: UserInfo }*/ = {
   admin: { businessName: "Admin", ...credentials.ADMIN },
-  customer1: { businessName: "Customer1", ...credentials.CUSTOMER1 },
-  customer: { businessName: "Customer2", ...credentials.CUSTOMER2 },
-};
-
-export const adminEndPoint = "admin";
-
-export const customerEndPoint = "account";
-
-// Comment const type to allow syntax error upon accessing non-exiting property
-export const apis /*: { [key: string]: ApiInfo }*/ = {
-  currentUser: {
-    interceptorName: "currentUser",
-    urlRegex: /users\/me$/,
-  },
-  yearsSales: {
-    interceptorName: "yearsSales",
-    urlRegex: /reports\/total-sales-of-years/,
-  },
-  dummyApi: {
-    interceptorName: "dummyInterceptorName",
-    urlRegex: /dummy regex/,
-    relativeUrl: (param: string) => `dynamic/${param}/dummyUrl`,
-  },
+  user1: { businessName: "User1", ...credentials.USER1 },
+  user2: { businessName: "User2", ...credentials.USER2 },
 };
