--- cypress/support/models/apiInfo.ts
@@ -1,5 +1,0 @@
-export type ApiInfo = {
-  readonly interceptorName: string;
-  readonly urlRegex: RegExp;
-  readonly relativeUrl?: (...args) => string;
-};
