+++ cypress/test/features/sepcialDummyFeature/special1.steps.ts
@@ -1,5 +1,5 @@
 import { Given } from "@badeball/cypress-cucumber-preprocessor";
 
-Given("Special2 dummy step", function () {
+Given("Special1 Dummy step", function () {
   cy.log("this is dummy a dummy step");
 });
